pru: ADC reader
Compute the watermark position based on the enabled channels (Eg, based
on the amount of data-samples in a 'set').

Allow changing the channels only when measurement is stopped to simplify
the buffer format. Eg, to ensure the buffer will contain samples from
only a currently enabled set of channels.

Write channel config to tx-buffer.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>