iio: bd79104-rpmsg: Add RPMSG => IIO bridge
The BD79104 ADC is able to measure 1KSPS when SPI is clocked at 20MHz.
Testing this with a Beaglebone Black Linux setup is hard.

In order to improve the throughput we can offload the SPI handling to
the PRU cores. This leaves a bit more time for the ARM to do data
processing. Thus the BD79104 is now accessed bu a PRU core, which reads
the data and fills it in a RPMSG buffer, generating an interrupt to ARM
when buffer is full.

Add a test driver which copies the ADC data from RPMSG buffer and
pushes it to IIO buffers.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>
---
This is a very crude code which was mainly used to see what I can
accomplish with hardware I have lying around. It seems that this type of
"data demuxing" on ARM at RPMSG IRQ takes around 200 uS. This is still way
too much. I could improve the situation in various ways, like completely
preparing the buffer on PRU side, dropping the timestamps or discarding
the RPMSG and delivering larger buffers to ARM.

Well, assuming the ADC data needs some further processing, the
Beaglebone Black's single-core ARM is likely to still be choking. Hence
this code has not been (and probably will not be) further polished.