mfd: bd71828: Support ROHM BD72720 and BD73900
The ROHM BD72720 and BD73900 are power management ICs which continue the
BD71828 family of PMICs. The BD79300 is, from the software
point-of-view, identical to the BD72720.

Similarly to the BD71815 and BD71828, these PMICs integrate regulators,
charger, RTC, clock gate and GPIOs.

The main difference to the earlier PMICs is that the BD72720 and BD79300
have two different I2C slave addresses. In addition to the registers
behind the 'main I2C address', charger (and to some extent LED) control
is done via registers behind a 'secondary I2C slave address'.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>