mfd: bd71828: Support ROHM BD72720
The ROHM BD72720 is a power management IC which continues the BD71828
family of PMICs. Similarly to the BD71815 and BD71828, the BD72720
integrates regulators, charger, RTC, clock gate and GPIOs.

The main difference to the earlier PMICs is that the BD72720 has two
different I2C slave addresses. In addition to the registers behind the
'main I2C address', most of the charger (and to some extent LED) control
is done via registers behind a 'secondary I2C slave address', 0x4c.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>

---
Revision history:
 v2 =>:
 - no changes

 RFCv1 => v2: (Mostly addressed comments from Lee and Andreas)
 - Use stacked regmaps to avoid platform data and the tango with
   multiple regmaps in the power-supply driver
 - Use regmap_reg_range()
 - make it clear bd72720_irq_type_base is an array
 - tab-out definitions in the bd72720 header
 - minor styling

Note: This patch depends on the series: "power: supply: add charger for
BD71828" by Andreas:
https://lore.kernel.org/all/20250918-bd71828-charger-v5-0-851164839c28@kemnade.info/

There are some new variants being planned. Most notably, the BD73900
should be almost identical to the BD72720 - for everything else except
the charger block.