dominion: Add function for form creation
The randomizer contains a form where user can input the extensions to
use, as well as the tuhina, tupina and kapita -preferences. Split this
code into own function, which can be used by the new randomizer.

NOTE: The new randomizer allows user to pick the tuhina, tupina and
capita-meters without a check-box. Value '0' is no longer 'remove cards
with tuhina/tupina/kapita' - but means "no preference, randomize without
weighing". Hence the form is different for old and new randomizer, and
the old randomizer still implements this form in index.php. Eg, old
randomizer does not use this function.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>