index.php: Avoid some warnings
We pass $exp and tuhina/tupina/kapita information variables to
functions whether the user gave them or not. This causes some warnings
about using unset variables.

Fix this by initializing these variables even when user has not given
them. Setting expansion ID to '0' works, as it is expected to be an
array when expansion is given. Also, initializing tuhina/tupina/kapita
to null will do the trick as we can use isset() to distinguish set/not
set info.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>