include: card: Add a helper to contain for card info
The old approach of doing card randomizing by weighed SQL RAND() and
ORDER BY has a few drawbacks.

Sorting entire table for randomizing may not be the best performing
option. I think this does not really matter with the amount of users
this has :) (Me and my little sister I suppose XD).

Bigger problem is that doing the weighed randomizer in SQL is getting
somewhat tricky. It is much more maintainable if weighing and
randomizing the cards is done in PHP side.

For this purpose, it is cleaner to encapsulate the card data in own
class. (I would prefer C-struct but .. meh).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>