dom_card_set: Add helper for final card data sets
When the randomizing is done in PHP side, it will be more lightweight
option to fetch first only a subset of card data for randomizing. Then,
when final set of cards is chosen, we can fetch full data for these
cards and finally show them to the audience. Furthermore, the cards are
grouped into 3 groups, cheap, mid-range and expensive cards.

Lets create a class for grouping and showing the cards.

NOTE: This is not really well done. The amount of cards in each group is
hardcoded here-and-there. Ideally this class should allow creating any
size of groups, and the group size should only be determined by the
amount of cards added to each 'set'.

This helper class will be used by the new randomizer.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>