index2: new randomizer
The old randomizer was a bit tricky to update. The randomizing was done
in the SQL, using ORDER BY (RAND() * weigh-options). This allowed the
randomizer to use single query, fetching only the information for
selected cards. As a donside, the weigh-options were now somewhat limited
and hard to change. (The SQL gets very hairy, very quickly).

In order to allow fine-tuning the weighing, and to support some future
changes (to be decided :] Don't hold your breath waiting), I decided to
change the randomizing to be done in PHP side.

As I didn't want to first read whole card database to memory for
randomizing, I had to increase the amount of queries.

Now we first query only the data required for randomizing. This got
quite a bit worse as I dropped the fixed cost grouping. (Old randomizer
always fetched 3 cards costing < 4, 3 cards costing 4 and finally, 4
cards costing more than 4). The main motivation was that when user plays
with 'Nousukausi' addon only, there won't be more than 3 cards costing
less than 4 in whole set. So, each randomizer turn picks the 3 same cards.

New randomizer first queries prizes for all cards in selected
expansions and then splits them in 3, roughly same size of cost-groups.
This means we added one large query getting the prize for all the cards.

Then we have one limited query / prize range, getting the card information
which is necessary for the randomizing. After randomizing is done, we
do one last query getting all the information for the selected cards.

So yes, this is quite a bit heavier than the old "do it all in SQL"
solution - although sorting the large tables for ORDER BY RAND() was not
exactly cheap either. Anyways, I hope this gets more maintainable.

The new randomizer has also minor change in UI. It is no longer
necessary to check the 'enable tuhina,tupina,capita,whatever'-box. Now
the default is to use XXX'o-meter value '0' - meaning no weighing is
done. Then the change in slider will immediately mean weighing is done
(either to favour or unfavour) to the selection. No checkbox needed.
(Well, 'nihilism' is a boolean option, having the checkbox selection).

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>