card.php: Fix randomizing for new expansions
The new expanisons may have NULL in database for the booleans like
'attack' or 'defence'. Thus we should not require them to be in a row
fetched from the database. Additionally, initialize booleans in card
class to '0' when setting is not found. This way check like

if ($card->attack)
is valid, and returns correct 'false' assumption if it is not explicitly
set to true.

Fix the new randomizer for new expansions which may have NULL flags in the
database, meaning 'false'.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>