index2: revert idea back to fixed cost buckets
The idea of dynamically splitting the cards to (roughly equally sized)
prize groups is not as good idea as it originally felt.

First of all, the new algorithm fails with some expansions, by not
having enough cards in the last prize group.

And more fundamentally, it may not be a great idea to randomize cards
which all cost 4 or more (which is possible with the dynamic prize
groups) - because it will probably make the first round(s) of game a
little boring because most of the players would get 3 and 4 coppers at
first rounds - resulting them to have no choise but to buy silver.

So, let's revert back to the original grouping:

3 Cheap cards (less than 4)
3 Mid range cards (exactly 4)
4 Expensive cards (more than 4)

It has been working rather well.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>