admin: Add 'tuhinaeditori'
The 'Tuhinafactor' value in the database is used to weigh cards which
increase chances of building an engine with a lot of actions. Assigning
the 'tuhinaweigh' is not so straightforward, and it may be these values
need to be altered when games are being played with new cards. Let's
create a simple tool for comparing and changing the 'tuhinafactor'.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>