footer: drop the visitor counter
I don't really like collecting information about the users. Furthermore,
loading scripts from external sites may not be fast.

It sure would be nice to know if this page is used - but the visitor
counter is not really a good way in these times... Times when all kinds
of bots and scrapers roam the web anyways. Knowing if an user is a real
human or not is not straightforward.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>