add_cards_admin: Fix the get_expansions() for admin
The an additional parameter needs to be passed to the get_expansions()
if all the expansions are queried. (Eg, expansions with the parsed_cards
data, not just expansions for which the cards are already added).

Admin panel needs to get all expansions so we can add card data for new
ones. So, get all the expansions.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>