card.php: Support table-prefixes
Idea was to allow the caller to pass the table-prefix used in SQL query,
so that fields with same name from different tables could be separated.

I am not really sure if this commit makes any sense. I don't think the
table prefixes are really used in fetched rows(?). It was already too
late when I hastly suffled (pun intended) this in.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>