Help deciding if a card is worth keeping
Add the information about the randomized cards so users have easier task
to decide if they wish to keep them, without having to find the card
itself. Additionally, this should help when doing a setup for the game,
randomizer should remind to pick the non-standard, expansion specific
stuff in play.

My very limited experience says that the curses and attacks, as well as
the added complexity are the main factors for some people when
considering if the card is good or not. So, lets try to illustrate
those qualities.

This is realized by using icons, which display an explanation when
pointed by mouse.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>