Help user picking expansions and me debugging
Picking the expansions is difficult because the checkboxes aren't
aligned with the expansion name. User needs to be careful to pick
correct checkbox, and it is quite inconvenient.

Force checkboxes to be aligned with the expansion name.

Furthermore, reading the HTML output by PHP is teodious due to the long
lines. This makes debugging cumbersome. Improve readability by adding
linefeeds to HTML generated in loops.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>