Document some improvement ideas
Adding the landmarks / events is a bit cumbersome, and adds quite a bit
of extra code. With some refactoring we might be able to use the
existing classes and avoid a lot of code. Unfortunately, I don't have
the energy for this right now - maybe later when my head hits the wall
with current approach :)

Add TODO items

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>