Simplify and clarify event deselection
The selection for the events and landmarks for an expansion will be
visible once the expansion is selected. Similarly, when the expansion is
deselected, the event and landmark selection disappears.

Disappearing however, does not deselect the events/landmarks. This feels
somewhat confusing and unexpected.

clean event and landmark checkboxes when related expansion is deselected.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>