dominion_common.php: Finnish + Improve sliders
I accidentally used English word for Landmark in UI. My intention is to
create a Finnish page - so let's replace the 'Landmark' with
'Maamerkit'. There already is an excellent english randomizer in the
web.

Additionally, the o-meter sliders got reset every time the randomizing
was done. This is not really optimal for users, who often repeat the
randomizing to replace some of the cards - but wish to keep the
preferences.

Initialize o-meter slider values to what was POST'ed.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>