index.php: escape html chars, fix landmark mobile row
The database may include characters which have special meaning in HTML.
Some of the database rows weren't filtered for these characters, which
makes the produced HTML to potentially broken.

Filter the database data through the htmlspecialchars() before sending
it to the browser.

Additionally, fix the landmark table for mobile view.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>