Add debt symbol
One thing requiring special setup is using debt. Currently, the cards
which can be purchased using debt are marked by having the currency type
appended to the prize. This, however is not very easy to spot, and it
does also take quite some space.

Furthermore, some cards may not cost debt - but may provide
functionality requiring debt. This means that having prize as 'debt' is
incorrect, while omitting the debt does not make it visible the setup
should include debt tokens.

We added debt in setup_extras in database, but let's additionally an
icon for debt here, which can be displayed instead of having the prize
type appended to the card prize.

Besides, the debt icon looks cool to me :)

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>