dominion_common.php: Sort expansions
Sort the expansions based on the order in which they have appeared, from
oldest to the newest. Keep the promo cards (extras) last.

We can't do sorting based on the IDs, because the "extras" were added
to the database prior to the "Rising Sun". Furthermore, we may get new
expansions.

Add new field to the database for the ordernumber and do sorting based
on it.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>