dominion_common.php: Get the event type name for expansion form from the database
When supporting different 'non buyable kingdom modifier cards' like
Landscapes, Ways, Projects, Traits and Prophecies, the type of the card
included in the expansion can no longer be hardcoded to the
'Landscapes'.

So, let's get it from the database.

The event_type table already contains the type of a card so it can be
displayed when card is picked by randomizer. This, however, is in
singular form, as it applies to a randomized card. (Eg "Maamerkki"
[landscape]) - whereas the set includes multiple cards and type should
be plural (Eg "Maamerkit" [landscapes]). For most of the english words
we could simply add 's' to the end - but not for all (like prophecy).
This, however, is not the case with the Finnish words, where the word
body may also change.

Add a new column for plural type names in the event_types table and
retrieve the name from it when expansions are queried.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>