dom_card_set: Fix prize buckets by ordering cards
The cards fetched by ID list will not be in the same order as the listed
IDs. So, we lose the relation to the 'set names' Eg card prizes.

Work around this by ordering the cards by Prize. This way we get the
cheapest cards first, and with the fixed number of cards in each prize
buckets, we know in which bucket a card belongs.

Signed-off-by: Matti Vaittinen <mazziesaccount@gmail.com>